refactor: move file download logic into SecretService
Extracts downloadFileSecret() from both web and API controllers into SecretService to eliminate duplication. Also clears file_size and file_mime_type on download alongside filepath/filename.
